+++ app/gui/src/project-view/components/GraphEditor/widgets/WidgetApplication.vue
@@ -8,9 +8,11 @@ import { entryMethodPointer } from '@/stores/suggestionDatabase/entry'
 import { Ast } from '@/util/ast'
 import { ArgumentApplication, ArgumentApplicationKey } from '@/util/callTree'
 import { computed } from 'vue'
+import { mapOrUndefined } from 'ydoc-shared/util/data/opt'
 
 const props = defineProps(widgetProps(widgetDefinition))
 const tree = injectWidgetTree()
+
 const application = computed(() => props.input[ArgumentApplicationKey])
 const graph = useGraphStore()
 
@@ -35,11 +37,9 @@ const appClass = computed(() => {
 })
 
 const operatorStyle = computed(() => {
-  if (
-    application.value.appTree instanceof Ast.OprApp ||
-    application.value.appTree instanceof Ast.PropertyAccess
-  ) {
-    const [_lhs, opr, rhs] = application.value.appTree.concreteChildren({
+  const appTree = application.value.appTree
+  if (appTree instanceof Ast.OprApp || appTree instanceof Ast.PropertyAccess) {
+    const [_lhs, opr, rhs] = appTree.concreteChildren({
       verbatim: true,
       indent: '',
     })
@@ -50,14 +50,18 @@ const operatorStyle = computed(() => {
   }
   return {}
 })
+
+const infixWidgetInput = computed(() =>
+  mapOrUndefined(application.value.infixOperator, WidgetInput.FromAst),
+)
+const showArgument = computed(() => tree.extended || !application.value.argument.hideByDefault)
+const argumentWidgetInput = computed(() => application.value.argument.toWidgetInput())
 </script>
 
 <script lang="ts">
 export const widgetDefinition = defineWidget(
   ArgumentApplicationKey,
-  {
-    priority: -20,
-  },
+  { priority: -20 },
   import.meta.hot,
 )
 </script>
@@ -65,15 +69,11 @@ export const widgetDefinition = defineWidget(
 <template>
   <div class="WidgetApplication" :class="appClass">
     <NodeWidget :input="targetMaybePort" :nest="application.isInnermost" />
-    <div v-if="application.infixOperator" class="infixOp" :style="operatorStyle">
-      <NodeWidget :input="WidgetInput.FromAst(application.infixOperator)" />
+    <div v-if="infixWidgetInput" class="infixOp" :style="operatorStyle">
+      <NodeWidget :input="infixWidgetInput" />
     </div>
     <SizeTransition width leftGap>
-      <NodeWidget
-        v-if="tree.extended || !application.argument.hideByDefault"
-        :input="application.argument.toWidgetInput()"
-        nest
-      />
+      <NodeWidget v-if="showArgument" :input="argumentWidgetInput" nest />
     </SizeTransition>
   </div>
 </template>